Started refactoring GDS CLI into utility files
